Refactor and enhance training center API
- Replaced empty controllers with fully implemented versions.
- Updated `ApplicationDbContext` to improve timestamp handling.
- Added `SeedData` for initial database population.
- Refactored `Enrollment` and `TrainingTrack` models.
- Implemented services for enrollments, instructors, payments, students, and tracks.
- Introduced new interfaces for service contracts.
- Added controllers for API endpoints with validation.
- Created DTOs for requests and responses.
- Standardized API responses with `ApiResponse<T>` and `PagedResult<T>`.
- Added `ReportService` for generating various reports.
- Updated enums for `EnrollmentStatus` and `PaymentMethod`.
- Created database migration for schema definition.
- Updated `Program.cs` to include JSON options and seed data.

+++ phase 03 - Real Backend Data Systems/Task 03 - Training Center Database API/Entities/Enrollment.cs
@@ -9,7 +9,7 @@ namespace Task_03___Training_Center_Database_API.Entities
         public EnrollmentStatus Status { get; set; }
         public DateTime EnrollmentDate { get; set; }
         public decimal ProgressPercentage { get; set; }
-        public decimal? FinalResult { get; set; }
+        public double? FinalResult { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
